feat: organizing the code
