+++ src/model/patient.model.ts
@@ -15,30 +15,3 @@ export interface IPatient {
 	behavior?: boolean; // little = false, much = true
 	behaviorScore?: number;
 }
-
-export interface IResponsePatients {
-	id: string;
-	name: string;
-	score: number;
-	behavior: string;
-}
-
-export interface IResponseDetailsPatientsDebug {
-	id: string;
-	name: string;
-	distance: number;
-	age: number;
-	acceptedOffers: number;
-	canceledOffers: number;
-	averageReplyTime: number;
-	score: {
-		age: number;
-		accepted: number;
-		canceled: number;
-		reply: number;
-		distance: number;
-		total: number;
-	};
-	behaviorScore?: number;
-	behavior?: string;
-}
